requirements.txt
numpy>=1.26.0
tensorflow>=2.15.0
tqdm>=4.65.0
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
mlflow>=2.9.0
tabulate>=0.9.0